requirements.txt
# HTTP requests library for API calls
requests>=2.31.0

# Data validation and parsing
pydantic>=2.5.0

# Environment variable management  
python-dotenv>=1.0.0

# Type hints support (included in Python 3.5+ but useful for older versions)
typing-extensions>=4.0.0